third_party/blink/web_tests/external/wpt/storage-access-api/storage-access-beyond-cookies.cookies.sub.https.window-expected.txt

This is a testharness.js-based test.
[FAIL] Verify StorageAccessAPIBeyondCookies for Cookies
  assert_equals: Storage Access API should be accessible and return first-party data expected "HasAccess for cookies" but got "First-party cookies should not be readable before handle is loaded."
Harness: the test ran to completion.
